CP-6096: Move existing iSCSI code to a base class
As a temporary step, move the existing code to a new base class
but symlink the existing ISCSISR file to the new base class.

This interim state has not been tested, and the next patch is required
to make sure the tests pass.  It is included only to show the rename
of the ISCSI base class in the patch series.
